+++ packages/cli-kit/src/analytics.ts
@@ -1,31 +1,35 @@
-/* eslint-disable @typescript-eslint/naming-convention */
 import * as environment from './environment.js'
 import {platformAndArch} from './os.js'
-import {exists as fileExists} from './file.js'
-import {join as joinPath, resolve} from './path.js'
+import {resolve} from './path.js'
 import {version as rubyVersion} from './node/ruby.js'
 import {content, debug, token} from './output.js'
 import constants from './constants.js'
-import {cliKitStore} from './store.js'
 import * as metadata from './metadata.js'
-import {publishEvent} from './monorail.js'
+import {publishEvent, MONORAIL_COMMAND_TOPIC} from './monorail.js'
 import {fanoutHooks} from './plugins.js'
+import {packageManagerUsedForCreating} from './node/node-package-manager.js'
 import {Interfaces} from '@oclif/core'
 
 interface StartOptions {
   command: string
   args: string[]
   currentTime?: number
+  commandClass?: Interfaces.Command.Class
 }
 
-export const start = ({command, args, currentTime = new Date().getTime()}: StartOptions) => {
-  metadata.addSensitive({
+export const start = async ({command, args, currentTime = new Date().getTime(), commandClass}: StartOptions) => {
+  await metadata.addSensitive(() => ({
     commandStartOptions: {
       startTime: currentTime,
       startCommand: command,
       startArgs: args,
     },
-  })
+  }))
+
+  await metadata.addPublic(() => ({
+    cmd_all_launcher: packageManagerUsedForCreating(),
+    cmd_all_plugin: commandClass?.plugin?.name,
+  }))
 }
 
 interface ReportEventOptions {
@@ -46,11 +50,11 @@ export async function reportEvent(options: ReportEventOptions) {
       // Nothing to log
       return
     }
-    if (environment.local.analyticsDisabled()) {
+    if (!environment.local.alwaysLogAnalytics() && environment.local.analyticsDisabled()) {
       debug(content`Skipping command analytics, payload: ${token.json(payload)}`)
       return
     }
-    const response = await publishEvent('app_cli3_command/1.0', payload.public, payload.sensitive)
+    const response = await publishEvent(MONORAIL_COMMAND_TOPIC, payload.public, payload.sensitive)
     if (response.type === 'error') {
       debug(response.message)
     }
@@ -78,27 +82,11 @@ const buildPayload = async ({config, errorMessage}: ReportEventOptions) => {
   if (pathFlagIndex >= 0) {
     directory = resolve(startArgs[pathFlagIndex + 1])
   }
-  const appInfo = cliKitStore().getAppInfo(directory)
 
-  const {platform, arch} = platformAndArch()
-
-  const rawPartnerId = appInfo?.orgId
-  let partnerIdAsInt: number | undefined
-  if (rawPartnerId !== undefined) {
-    partnerIdAsInt = parseInt(rawPartnerId, 10)
-    if (isNaN(partnerIdAsInt)) {
-      partnerIdAsInt = undefined
-    }
-  }
-
-  const publicPluginData = await fanoutHooks(config, 'public_command_metadata', {})
+  const {'@shopify/app': appPublic, ...otherPluginsPublic} = await fanoutHooks(config, 'public_command_metadata', {})
   const sensitivePluginData = await fanoutHooks(config, 'sensitive_command_metadata', {})
 
-  const appSpecific = {
-    project_type: await getProjectType(joinPath(directory, 'web')),
-    api_key: appInfo?.appId,
-    partner_id: partnerIdAsInt,
-  }
+  const {platform, arch, ...environmentData} = getEnvironmentData(config)
 
   return {
     public: {
@@ -112,33 +100,40 @@ const buildPayload = async ({config, errorMessage}: ReportEventOptions) => {
       ruby_version: (await rubyVersion()) || '',
       node_version: process.version.replace('v', ''),
       is_employee: await environment.local.isShopify(),
-      ...appSpecific,
+      ...environmentData,
+      ...appPublic,
+      ...metadata.getAllPublic(),
     },
     sensitive: {
       args: startArgs.join(' '),
       error_message: errorMessage,
       metadata: JSON.stringify({
         ...sensitiveMetadata,
-        extraPublic: publicPluginData,
+        extraPublic: {
+          ...otherPluginsPublic,
+        },
         extraSensitive: sensitivePluginData,
       }),
     },
   }
 }
 
-export type ProjectType = 'node' | 'php' | 'ruby' | undefined
+function getEnvironmentData(config: Interfaces.Config) {
+  const ciPlatform = environment.local.ciPlatform()
 
-export async function getProjectType(directory: string): Promise<ProjectType> {
-  const nodeConfigFile = joinPath(directory, 'package.json')
-  const rubyConfigFile = joinPath(directory, 'Gemfile')
-  const phpConfigFile = joinPath(directory, 'composer.json')
+  const pluginNames = config.plugins
+    .map((plugin) => plugin.name)
+    .sort()
+    .filter((plugin) => !plugin.startsWith('@oclif/'))
+  const shopifyPlugins = pluginNames.filter((plugin) => plugin.startsWith('@shopify/'))
 
-  if (await fileExists(nodeConfigFile)) {
-    return 'node'
-  } else if (await fileExists(rubyConfigFile)) {
-    return 'ruby'
-  } else if (await fileExists(phpConfigFile)) {
-    return 'php'
+  return {
+    ...platformAndArch(),
+    env_ci: ciPlatform.isCI,
+    env_ci_platform: ciPlatform.name,
+    env_plugin_installed_any_custom: pluginNames.length !== shopifyPlugins.length,
+    env_plugin_installed_shopify: JSON.stringify(shopifyPlugins),
+    env_shell: config.shell,
+    env_web_ide: environment.local.webIDEPlatform(),
   }
-  return undefined
 }