+++ packages/cli-kit/src/plugins.ts
@@ -1,6 +1,10 @@
 import {join, pathToFileURL} from './path.js'
 import {debug, content} from './output.js'
-import {Plugin, Config, Hooks} from '@oclif/core/lib/interfaces'
+import {JsonMap} from './json.js'
+import {PickByPrefix} from './typing/pick-by-prefix.js'
+import {MonorailEventPublic} from './monorail.js'
+import {HookReturnPerTunnelPlugin} from './plugins/tunnel.js'
+import {Interfaces} from '@oclif/core'
 
 const TUNNEL_PLUGINS = ['@shopify/plugin-ngrok']
 
@@ -12,7 +16,7 @@ interface TunnelStartOptions {
   port: number
 }
 
-export async function lookupTunnelPlugin(plugins: Plugin[]): Promise<TunnelPlugin | undefined> {
+export async function lookupTunnelPlugin(plugins: Interfaces.Plugin[]): Promise<TunnelPlugin | undefined> {
   debug(content`Looking up the Ngrok tunnel plugin...`)
   const tunnelPlugin = plugins.find((plugin) => TUNNEL_PLUGINS.includes(plugin.name))
   if (!tunnelPlugin) return undefined
@@ -25,12 +29,46 @@ export async function lookupTunnelPlugin(plugins: Plugin[]): Promise<TunnelPlugi
  *
  * Responses are organised into a dictionary, keyed by plug-in name. Only plug-ins that have hooks registered for the given event, and the hooks were run successfully, are included.
  */
-export async function fanoutHooks<T extends keyof Hooks, TResult = Hooks[T]['return']>(
-  config: Config,
-  event: T,
-  options: Hooks[T]['options'],
+export async function fanoutHooks<TPluginMap extends HookReturnsPerPlugin, TEvent extends string & keyof TPluginMap>(
+  config: Interfaces.Config,
+  event: TEvent,
+  options: TPluginMap[typeof event]['options'],
   timeout?: number,
-): Promise<{[pluginName: string]: TResult}> {
+): Promise<Partial<TPluginMap[typeof event]['pluginReturns']>> {
   const res = await config.runHook(event, options, timeout)
-  return Object.fromEntries(res.successes.map(({result, plugin}) => [plugin.name, result]))
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  return Object.fromEntries(res.successes.map(({result, plugin}) => [plugin.name, result])) as any
 }
+
+type AppSpecificMonorailFields = PickByPrefix<MonorailEventPublic, 'app_', 'project_type' | 'api_key' | 'partner_id'> &
+  PickByPrefix<MonorailEventPublic, 'cmd_extensions_'> &
+  PickByPrefix<MonorailEventPublic, 'cmd_scaffold_'>
+
+interface HookReturnsPerPlugin extends HookReturnPerTunnelPlugin {
+  public_command_metadata: {
+    options: {[key: string]: never}
+    pluginReturns: {
+      '@shopify/app': Partial<AppSpecificMonorailFields>
+      [pluginName: string]: JsonMap
+    }
+  }
+  [hookName: string]: {
+    options: {[key: string]: unknown}
+    pluginReturns: {[key: string]: JsonMap}
+  }
+}
+
+export type PluginReturnsForHook<
+  TEvent extends keyof TPluginMap,
+  TPluginName extends keyof TPluginMap[TEvent]['pluginReturns'],
+  TPluginMap extends HookReturnsPerPlugin = HookReturnsPerPlugin,
+> = TPluginMap[TEvent]['pluginReturns'][TPluginName]
+
+export type FanoutHookFunction<
+  TEvent extends keyof TPluginMap = string,
+  TPluginName extends keyof TPluginMap[TEvent]['pluginReturns'] = string,
+  TPluginMap extends HookReturnsPerPlugin = HookReturnsPerPlugin,
+> = (
+  this: Interfaces.Hook.Context,
+  options: TPluginMap[TEvent]['options'] & {config: Interfaces.Config},
+) => Promise<PluginReturnsForHook<TEvent, TPluginName, TPluginMap>>