+++ packages/create-app/src/prompts/init.ts
@@ -3,6 +3,7 @@ import {haiku, ui} from '@shopify/cli-kit'
 interface InitOptions {
   name?: string
   template?: string
+  directory: string
 }
 
 interface InitOutput {
@@ -20,7 +21,7 @@ export const templateURLMap = {
 
 const init = async (options: InitOptions, prompt = ui.prompt): Promise<InitOutput> => {
   const defaults = {
-    name: haiku.generate('app'),
+    name: await haiku.generate({suffix: 'app', directory: options.directory}),
     template: templateURLMap.node,
   } as const
 
@@ -45,10 +46,16 @@ const init = async (options: InitOptions, prompt = ui.prompt): Promise<InitOutpu
   }
 
   if (!options.template && Object.keys(templateURLMap).length > 1) {
+    const templateList = Object.keys(templateURLMap).map((key) => {
+      return {
+        name: key,
+        value: key,
+      }
+    })
     questions.push({
       type: 'select',
       name: 'template',
-      choices: Object.keys(templateURLMap),
+      choices: templateList,
       message: 'Which template would you like to use?',
       default: defaults.template,
     })