+++ packages/create-app/src/services/init.ts
@@ -58,9 +58,7 @@ async function init(options: InitOptions) {
               task: async (_, task) => {
                 task.title = 'Parsing liquid'
                 await template.recursiveDirectoryCopy(templatePathDir, templateScaffoldDir, {
-                  // eslint-disable-next-line @typescript-eslint/naming-convention
                   dependency_manager: packageManager,
-                  // eslint-disable-next-line @typescript-eslint/naming-convention
                   app_name: options.name,
                 })
 
@@ -164,7 +162,8 @@ function inferPackageManager(optionsPackageManager: string | undefined): Package
   if (optionsPackageManager && packageManager.includes(optionsPackageManager as PackageManager)) {
     return optionsPackageManager as PackageManager
   }
-  return packageManagerUsedForCreating()
+  const usedPackageManager = packageManagerUsedForCreating()
+  return usedPackageManager === 'unknown' ? 'npm' : usedPackageManager
 }
 
 async function ensureAppDirectoryIsAvailable(directory: string, name: string): Promise<void> {