+++ packages/app/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @shopify/app
 
+## 3.7.1
+
+### Minor Changes
+
+- bba213f9: Shopify Functions identifiers have been changed to ULIDs. Existing projects will be updated during the next deploy.
+
+### Patch Changes
+
+- Updated dependencies [18717ad5]
+- Updated dependencies [29f46e8f]
+- Updated dependencies [bba213f9]
+  - @shopify/cli-kit@3.7.1
+  - @shopify/shopify-cli-extensions@3.7.1
+
+## 3.6.2
+
+### Patch Changes
+
+- Updated dependencies [59d56a40]
+  - @shopify/cli-kit@3.6.2
+
+## 3.6.1
+
+### Patch Changes
+
+- c33a6dcf: Bug-fix for running embedded apps without a backend
+  - @shopify/cli-kit@3.6.1
+
+## 3.6.0
+
+### Minor Changes
+
+- 1d438b3c: Support for new internal extension type identifiers
+
+### Patch Changes
+
+- d9351df4: Use a custom list of safe nouns and adjectives to generate default names for apps and extensions
+- 858e3faf: Human facing internal extension type ids
+- ffd33ec1: Fix logs when reusing dev values
+- Updated dependencies [073e514c]
+- Updated dependencies [d9351df4]
+  - @shopify/cli-kit@3.6.0
+
 ## 3.5.0
 
 ### Minor Changes